move BuildInDir to testutils
